refactor(frontend/message): 💄 message components
